feat: 여행 상세 조회 시, 여정, 댓글, 좋아요 수 까지 조회

+++ src/main/java/kr/co/fastcampus/travel/domain/trip/service/dto/response/TripItineraryInfoDto.java
@@ -14,7 +14,7 @@ public record TripItineraryInfoDto(
     LocalDate startDate,
     LocalDate endDate,
     boolean isForeign,
-    Long likeCount,
+    int likeCount,
     List<ItineraryResponse> itineraries
 ) {
